Use test containers to facilitate running various integration tests (#1931)
* Unignore Rider files to share them (as they should be) and add docker configuration to be able to run integration tests more easily

* Some cleanup

* Use testcontainers to start docker containers when reunning intergration tests, for : Postgresql, MySql, Oracle, SQLServer and Firebird (not tested)

* Restore gitignore

* Fix failing integration tests

* Add possibility to override appsettings with env vars to enable local containers

* Fix the Firebird connection string, and add it to the containers list

* Use the same version as the other packages

* Update Testcontainers dependencies

* Add missing dependabot dependencies

* Change Firebird tests to use the original connection string with Docker
Added DB2 docker support
Fixed CanUseRawSqlInUpdateAndDelete tests
Fixed Firebird integration tests with page size too small errors
Ignore Oracle native tests when no native client available

---------

+++ test/FluentMigrator.Tests/Integration/Processors/Firebird/TemporaryDatabase.cs
@@ -32,6 +32,18 @@ namespace FluentMigrator.Tests.Integration.Processors.Firebird
             if (!connectionOptions.IsEnabled)
                 Assert.Ignore();
 
+            if (connectionOptions.ContainerEnabled)
+            {
+                _connectionString = new Lazy<string>(() =>
+                {
+                    var csb = new FbConnectionStringBuilder(connectionOptions.ConnectionString);
+
+                    return prober.CreateDb(csb);
+                });
+
+                return;
+            }
+
             DbFileName = Path.GetTempFileName();
             _connectionString = new Lazy<string>(() =>
             {
@@ -53,7 +65,12 @@ namespace FluentMigrator.Tests.Integration.Processors.Firebird
         {
             try
             {
-                File.Delete(DbFileName);
+                if (!string.IsNullOrEmpty(DbFileName))
+                {
+                    File.Delete(DbFileName);
+                }
+
+                FbDatabase.DropDatabase(ConnectionString);
             }
             catch
             {
